requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
aiosqlite==0.20.0
python-dotenv==1.0.1
razorpay==1.4.2
fastapi==0.115.0
uvicorn==0.30.6
